renamed files from using camel case to kebab case
