+++ assets/js/components/hangman-image.js
