+++ assets/js/components/increase-score.js
